handler all object serializing in Serializer
