+++ src/util/objects.ts
@@ -23,23 +23,6 @@ export const arrayify = <T>(input: T): ToArray<T> => (
     Array.isArray(input) ? input : [input]
 ) as ToArray<T>;
 
-export const serializeObject = (obj: any) => (
-    JSON.stringify(
-        obj,
-        (key, val) => (
-            val && (
-                val instanceof RegExp ||
-                typeof val === "function"
-            ) ? val.toString() : val
-        ),
-    )
-);
-
-export const objectsAreEqual = (obj1: any, obj2: any) => {
-    obj1 = serializeObject(obj1);
-    obj2 = serializeObject(obj2);
-    return obj1 === obj2;
-};
 
 export const removeFromArray = <T>(arr: T[], val: T, firstOnly = false) => {
     for (let idx = arr.length - 1; idx >= 0; --idx) {
